third_party/WebKit/LayoutTests/http/tests/security/ftp-subresource-blocked.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test(t => {
    var i = document.createElement('img');
    i.onload = t.unreached_func('The image should not load.');
    i.onerror = t.step_func_done(e => {
      assert_equals(i.naturalWidth, 0);
      assert_equals(i.naturalHeight, 0);
    });

    i.src = 'ftp://example.test/image.png';
  }, "'ftp:' subresource requests fail.");
</script>
